Fixed issue #201
The client handles a failed request that required authentication by
re-executing the original call with basic auth. When an exception occurs
in the service after the client has been authenticated, the client still
delivers an "unauthorized" exception. This is incorrect, since by the
time you receive the exception the client has been authenticated and the
actual exception is something else.