service client methods to help download bytes
Added pair of methods that make it much more convenient to download
bytes from a service resource. The request DTO can be used to determine
the resource URL. Also, the service client's request and response
filters are observed. This is helpful, for example, when custom headers
sent to the server.