Add models and responseClass to SwaggerApiService
The Models property in ResourceResponse, and the ResponseClass property
in MethodOperation, allow the Swagger UI to display schemas and sample
JSON for complex request/response types. The model data is generated by
reflection on request and response DTOs and inspection of
ApiMember/Description/SwaggerAllowableValues attributes. To use this
feature, define one ApiMember attribute with ParameterType = "body" and
DataType = "<class name of your request DTO>", and mark the request DTO
class with the IReturn<ResponseDTOClassName> interface. Also, ensure
that the Swagger UI js/css/html is up to date.