Allow ApiMember attributes to be applied by verb and/or route
This allows different Swagger Models to be emitted for the same DTO
based on Route and/or Verb. Supports a use-case where the same DTO may
be used for both a PUT & POST but only certain fields on the DTO may be
applicable to one verb not both. Providing more granular api docs.