Proposing an extension to Default attribute
Rationale: When using the Default attribute, especially with arguments such
as OrmLiteVariables.SystemUtc, it is frequently true that you want the default
value set on an update. Currently the only way to do that in OrmLite is to
use custom SQL.

You may remember https://github.com/ServiceStack/ServiceStack/pull/1005 which
you rejected in favor of the Default attribute. So we are using the Default
attribute and finding that it needs update functionality. Hence this PR and if
you approve, the actual work to implement it in OrmLite.

The "UseOnUpdate=true" extension would be implemented by each dialect provider
by either generating SQL to set the column value on an update operation or by
creating a trigger which does it. For an UpdateOnly operation, any UseOnUpdate
columns would only be updated if the onlyFields expression includes them. For
an UpdateNonDefaults operation UseOnUpdate columns would not be updated.