Abstract missing type generator
Added an interface for the missing type generator in AutoQueryFeature.
This will give the user more control of the types that are generated for
the auto query feature.  My team needed a way to add custom attributes
to the generated types (specifically Authenticate).  Adding an interface
and base implementation with a hook to add custom attributes allows us
to decorate our request DTOs and override the AddCustomAttributes method
on the base IAutoQueryMissingTypeGenerator implementation.