Handle disconnected ServerSentEvent subscriptions
Currently if a client disconnects while a channel is being notified, it
causes an error to be thrown when attempting to SetResult on the
TaskCompletionSource. This patch will prevent setting the result on the
task if it's already completed. It will also prevent attempting to
notify a subscription where the underlying connection is already closed
(currently the notify calls publish which throws an exception which then
unsubscribes the current subscription - this will prevent the exception,
and simply unsubscribe.)