Preserve port from loadbalancer/proxy in front of ServiceStack
When ServiceStack is running behind a loadbalancer or proxy that correctly sets the header X-Forwarded-Port,
then the base url should change to reflect the proxy.
This allows a setup where ServiceStack listens to a specific port, but that port is only exposed to the
loadbalancer/proxy, as in the scenario:
http://my.domain/ -> LoadBalancer -> http://a.b.c.d:x/
where access to, for instance http://my.domain/ should have working links.