[netcore] System.Threading.ThreadPool package downgrade warning (#1068)
When I run dotnet restore on ServiceStack.Core, I get the following warning:

warn : Detected package downgrade: System.Threading.ThreadPool from 4.0.10 to 4.0.1 

So, this PR is to bump the version. I didn't investigate whether or not there was a reason for picking 4.0.1, but this was the only dependency throwing the warning, so I assumed it was a typo.