Distinguish formData and body params in OpenAPI
Swagger 2.0 specification does not allow to specify two "consumes"
content-type for the same operation with different parameters. This is
allowed only in Swagger 3.0 spec
https://github.com/OAI/OpenAPI-Specification/blob/OpenAPI.next/versions/3.0.md#support-for-x-www-form-urlencoded-request-bodies

So operation `consumes` content-type is distunguished based on `DisableAutoDtoInBodyParam`
setting. If this setting is `false` then we use `application/json` for
`POST` and `PUT` otherwise we use `application/x-www-form-urlencoded`