Added ReturnAttribute for use in OrmLite
Added the ReturnAttribute to implement in OrmLite.
Properties marked with this attribute will be included in the RETURNING (FireBird and PostgreSql) or OUTPUT (SqlServer) clause when building sql sentences for INSERT or UPDATE operations.

This allows to use SEQUENCES in SqlServer for Id columns instead of IDENTITY.
Also and not less useful, allows to get back field values that might be updated in database triggers when doing the same operations.