Fix memory leak: After disposing ServiceStackHost is kept in memory
Even after ServiceStackHost has been disposed, the static event TaskScheduler.UnobservedTaskException holds a reference to the object and thus it is kept in memory throughout the lifetime of the application. This creates a problem with systems that have "soft" restarts where the app domain is never disposed, and ServiceStack is reinitialized multiple times.