Refresh BearerToken properly in JsonHttpClient (#1214)
JsonHttpClient sends a refresh request if `RefreshToken` is set, but
while processing the refresh, the underlying `HttpClient` is not updated,
so subsequent requests with the same JsonHttpClient will again
need to refresh the out-of-date `BearerToken`.  The update to
`BearerToken` is never reflected in `HttpClient`.  Note that this only
happens with token-only authorization.  Cookie and Basic auth properly
update the `HttpClient`.