Update FirebirdOrmLiteDialectProvider to include ToTableNamesStatement
Allows for scaffolding crud endpoints using Autoquery.

N.B: It trims whitespace from the table names since RDB$RELATION_NAME is CHAR(31) which pads the string with spaces.
This will cause scaffolding to fail on tables like '__EFMigrationsHistory'.

This was just a quick fix I had to implement while trying to test ServiceStack today. Might be a better way to accomplish this whith a better understanding of the framework.